Initial commit
Going ahead and getting this pushed so I can leave it alone for now, and so my friend can use this as a reference. Included a basic README and gitignore for UWP projects that might need some fine-tuning. A lot of refactoring will probably happen soon but oh well :thumbsup: